+++ blueprints/__init__.py
@@ -1,0 +1,2 @@
+from blueprints.backend import backend
+from blueprints.frontend import frontend
