+++ templates/completion.html
@@ -40,6 +40,6 @@
     <p>Your task has been completed, and CSV files have been created.</p>
 
     <!-- Button to navigate to the front page for a new search -->
-    <button onclick="window.location.href='{{ url_for('index.html') }}'">Search Again</button>
+    <button onclick="window.location.href='/'">Search Again</button>
 </body>
 </html>
